adding stub for new tree behavior implementation
The tree behavior itself is a meta-behavior which implements no
functionality itself, but only loads an implemenation-specific behavior.

Since findThreaded is an example of tree-like behavior - moving that to
the AdjacentList behavior
